requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
python-dotenv==1.0.0
google-genai>=0.2.0
youtube-transcript-api==1.2.2
Werkzeug==2.3.7
SQLAlchemy>=2.0.43
psycopg2-binary>=2.9.9
google-api-python-client==2.108.0
requests>=2.31.0
